+++ docs/list.json
@@ -347,7 +347,7 @@
 				"DecalGeometry": "examples/en/geometries/DecalGeometry",
 				"ParametricGeometry": "examples/en/geometries/ParametricGeometry",
 				"TextGeometry": "examples/en/geometries/TextGeometry",
-				"SDFGeometry": "examples/en/geometries/SDFGeometry"
+				"SDFGeometryGenerator": "examples/en/geometries/SDFGeometryGenerator"
 			},
 
 			"Helpers": {
